+++ src/libc/phapi/bitmap.c
@@ -4,17 +4,17 @@
 #include <phapi.h>
 #include <sys/cdefs.h>
 
-void use_bit(uint32_t *bitmap, uint32_t bit)
+void use_bit(size_t *bitmap, size_t bit)
 {
     bitmap[BIT_INDEX(bit)] |= (0x1 << BIT_OFFSET(bit));
 }
 
-void free_bit(uint32_t *bitmap, uint32_t bit)
+void free_bit(size_t *bitmap, size_t bit)
 {
     bitmap[BIT_INDEX(bit)] &= ~(0x1 << BIT_OFFSET(bit)); 
 }
 
-uint32_t test_bit(uint32_t *bitmap, uint32_t bit)
+size_t test_bit(size_t *bitmap, size_t bit)
 {
     return !(bitmap[BIT_INDEX(bit)] & (0x1 << BIT_OFFSET(bit)));
 }
